requirements.txt
numpy
Pillow
scikit-image
scipy
tensorflow
